+++ AI-Chatbot/screens/chat/ChatScreen/styles.js
@@ -1,4 +1,4 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet, Platform } from 'react-native';
 import { theme } from '../../../styles/theme';
 
 export default StyleSheet.create({
